requirements.txt
# Core dependencies
numpy>=1.21.0
scikit-learn>=0.24.2
faiss-cpu>=1.7.2  # Use faiss-gpu for GPU support
mistralai>=0.0.7

# API and Web Server
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
starlette>=0.14.2

# Security
cryptography>=3.4.7
PyJWT>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-multipart>=0.0.5

# Authentication
authlib>=0.15.4
oauth2client>=4.1.3
oauthlib>=3.1.1

# Rate Limiting and Caching
redis>=4.0.2
aioredis>=2.0.0
limits>=1.5.1

# Monitoring and Logging
prometheus-client>=0.11.0
python-json-logger>=2.0.2
structlog>=21.1.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.18.2

# Development
black>=21.7b0
isort>=5.9.3
mypy>=0.910
flake8>=3.9.2
